+++ frontend/openchat-agent/src/services/openchatAgent.ts
@@ -3931,8 +3931,8 @@ export class OpenChatAgent extends EventTarget {
         return this.userClient.setPinNumber(verification, newPin);
     }
 
-    claimDailyChit(): Promise<ClaimDailyChitResponse> {
-        return this.userClient.claimDailyChit();
+    claimDailyChit(utcOffsetMins: number | undefined): Promise<ClaimDailyChitResponse> {
+        return this.userClient.claimDailyChit(utcOffsetMins);
     }
 
     chitLeaderboard(): Promise<ChitLeaderboardResponse> {